# openapi/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "Microservice to manage PagoPA Backoffice",
    "termsOfService" : "https://www.pagopa.gov.it/",
    "title" : "SelfCare Backoffice",
    "version" : "1.4.5"
  },
  "servers" : [ {
    "url" : "http://localhost:8080"
  }, {
    "url" : "https://{host}{basePath}",
    "variables" : {
      "basePath" : {
        "default" : "/backoffice/v1"
      },
      "host" : {
        "default" : "api.dev.platform.pagopa.it",
        "enum" : [ "api.dev.platform.pagopa.it", "api.uat.platform.pagopa.it", "api.platform.pagopa.it" ]
      }
    }
  } ],
  "paths" : {
...
    "/bundles/payment-service-providers/{psp-code}" : {
      "get" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y |  | N\n",
        "operationId" : "getBundlesByPSP",
        "parameters" : [ {
          "description" : "Number of elements on one page. Default = 50",
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 50
          }
        }, {
          "description" : "Commissional bundle's type",
          "in" : "query",
          "name" : "bundle-type",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "GLOBAL", "PUBLIC", "PRIVATE" ]
            }
          }
        }, {
          "description" : "Page number. Page value starts from 0",
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "description" : "Fiscal code of the payment service provider",
          "in" : "path",
          "name" : "psp-code",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Commissional bundle's name",
          "in" : "query",
          "name" : "name",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Bundles"
                }
              }
            },
            "description" : "OK",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "Get a paginated list of bundles related to PSP",
        "tags" : [ "Commission bundles" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ],
      "post" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y |  | N\n",
        "operationId" : "createBundle",
        "parameters" : [ {
          "description" : "Fiscal code of the payment service provider",
          "in" : "path",
          "name" : "psp-code",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BundleCreate"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "Created",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "Create a commissional bundle related to PSP",
        "tags" : [ "Commission bundles" ]
      }
    },
    "/bundles/payment-types" : {
      "get" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y |  | N\n",
        "operationId" : "getBundlesPaymentTypes",
        "parameters" : [ {
          "description" : "Number of elements on one page. Default = 50",
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 50
          }
        }, {
          "description" : "Page number. Page value starts from 0",
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BundlePaymentTypes"
                }
              }
            },
            "description" : "OK",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "Get a paginated list of payment types",
        "tags" : [ "Commission bundles" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ]
    },
    "/bundles/touchpoints" : {
      "get" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y |  | N\n",
        "operationId" : "getTouchpoints",
        "parameters" : [ {
          "description" : "Number of elements on one page. Default = 10",
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10
          }
        }, {
          "description" : "Page number. Page value starts from 0",
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Touchpoints"
                }
              }
            },
            "description" : "OK",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "Get a paginated list of touchpoints",
        "tags" : [ "Commission bundles" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ]
    },
...
    "/payment-service-providers/{psp-code}" : {
      "get" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y |  | N\n",
        "operationId" : "getBrokerAndPspDetails",
        "parameters" : [ {
          "description" : "PSP code",
          "in" : "path",
          "name" : "psp-code",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BrokerOrPspDetailsResource"
                }
              }
            },
            "description" : "OK",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "Get payment service provider's details",
        "tags" : [ "Payment Service Providers" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ],
      "put" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y |  | N\n",
        "operationId" : "updatePSP",
        "parameters" : [ {
          "description" : "code of the Payment Service Provider",
          "in" : "path",
          "name" : "psp-code",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PaymentServiceProviderDetailsDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaymentServiceProviderDetailsResource"
                }
              }
            },
            "description" : "OK",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "Update payment service provider",
        "tags" : [ "Payment Service Providers" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "BrokerOrPspDetailsResource" : {
        "type" : "object",
        "properties" : {
          "brokerPspDetailsResource" : {
            "$ref" : "#/components/schemas/BrokerPspDetailsResource"
          },
          "paymentServiceProviderDetailsResource" : {
            "$ref" : "#/components/schemas/PaymentServiceProviderDetailsResource"
          }
        }
      },
...
      "BrokerPspDetailsResource" : {
        "required" : [ "broker_psp_code", "description", "enabled", "extended_fault_bean" ],
        "type" : "object",
        "properties" : {
          "broker_psp_code" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "enabled" : {
            "type" : "boolean"
          },
          "extended_fault_bean" : {
            "type" : "boolean"
          }
        }
      },
...
      "Bundle" : {
        "type" : "object",
        "properties" : {
          "description" : {
            "type" : "string"
          },
          "id_bundle" : {
            "type" : "string"
          },
          "id_ci_bundle" : {
            "type" : "string"
          },
          "id_psp" : {
            "type" : "string"
          },
          "inserted_date" : {
            "type" : "string",
            "format" : "date-time"
          },
          "last_updated_date" : {
            "type" : "string",
            "format" : "date-time"
          },
          "max_payment_amount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "min_payment_amount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "name" : {
            "type" : "string"
          },
          "payment_amount" : {
            "type" : "integer",
            "description" : "The fees of this bundle",
            "format" : "int32"
          },
          "payment_type" : {
            "type" : "string"
          },
          "touchpoint" : {
            "type" : "string"
          },
          "transfer_category_list" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "type" : {
            "type" : "string",
            "enum" : [ "GLOBAL", "PUBLIC", "PRIVATE" ]
          },
          "validity_date_from" : {
            "type" : "string",
            "format" : "date"
          },
          "validity_date_to" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
      "BundleCreate" : {
        "required" : [ "abi", "id_broker_psp", "id_channel", "psp_business_name" ],
        "type" : "object",
        "properties" : {
          "abi" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "digital_stamp" : {
            "type" : "boolean"
          },
          "digital_stamp_restriction" : {
            "type" : "boolean"
          },
          "id_broker_psp" : {
            "type" : "string"
          },
          "id_cdi" : {
            "type" : "string"
          },
          "id_channel" : {
            "type" : "string"
          },
          "max_payment_amount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "min_payment_amount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "name" : {
            "type" : "string"
          },
          "payment_amount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "payment_type" : {
            "type" : "string"
          },
          "psp_business_name" : {
            "type" : "string"
          },
          "touchpoint" : {
            "type" : "string"
          },
          "transfer_categories" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "type" : {
            "type" : "string",
            "enum" : [ "GLOBAL", "PUBLIC", "PRIVATE" ]
          },
          "validity_date_from" : {
            "type" : "string",
            "format" : "date"
          },
          "validity_date_to" : {
            "type" : "string",
            "format" : "date"
          }
        },
        "description" : "Commissional bundle related to PSP to be created"
      },
      "BundlePaymentType" : {
        "type" : "object",
        "properties" : {
          "created_date" : {
            "type" : "string",
            "format" : "date-time"
          },
          "description" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        }
      },
      "BundlePaymentTypes" : {
        "required" : [ "page_info", "payment_types" ],
        "type" : "object",
        "properties" : {
          "page_info" : {
            "$ref" : "#/components/schemas/PageInfo"
          },
          "payment_types" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/BundlePaymentType"
            }
          }
        }
      },
      "Bundles" : {
        "type" : "object",
        "properties" : {
          "bundles" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Bundle"
            }
          },
          "page_info" : {
            "$ref" : "#/components/schemas/PageInfo"
          }
        }
      },
...
      "PageInfo" : {
        "type" : "object",
        "properties" : {
          "items_found" : {
            "type" : "integer",
            "format" : "int32"
          },
          "limit" : {
            "type" : "integer",
            "format" : "int32"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "total_items" : {
            "type" : "integer",
            "format" : "int64"
          },
          "total_pages" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "PaymentServiceProviderDetailsDto" : {
        "required" : [ "business_name", "enabled", "psp_code" ],
        "type" : "object",
        "properties" : {
          "abi" : {
            "type" : "string"
          },
          "agid_psp" : {
            "type" : "boolean"
          },
          "bic" : {
            "type" : "string"
          },
          "business_name" : {
            "type" : "string"
          },
          "enabled" : {
            "type" : "boolean"
          },
          "my_bank_code" : {
            "type" : "string"
          },
          "psp_code" : {
            "type" : "string"
          },
          "stamp" : {
            "type" : "boolean"
          },
          "tax_code" : {
            "type" : "string"
          },
          "vat_number" : {
            "type" : "string"
          }
        }
      },
      "PaymentServiceProviderDetailsResource" : {
        "required" : [ "business_name", "enabled", "psp_code" ],
        "type" : "object",
        "properties" : {
          "abi" : {
            "type" : "string"
          },
          "agid_psp" : {
            "type" : "boolean"
          },
          "bic" : {
            "type" : "string"
          },
          "business_name" : {
            "type" : "string"
          },
          "enabled" : {
            "type" : "boolean"
          },
          "my_bank_code" : {
            "type" : "string"
          },
          "psp_code" : {
            "type" : "string"
          },
          "stamp" : {
            "type" : "boolean"
          },
          "tax_code" : {
            "type" : "string"
          },
          "vat_number" : {
            "type" : "string"
          }
        }
      },
...
      "Touchpoint" : {
        "type" : "object",
        "properties" : {
          "created_date" : {
            "type" : "string",
            "format" : "date-time"
          },
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        }
      },
      "Touchpoints" : {
        "type" : "object",
        "properties" : {
          "pageInfo" : {
            "$ref" : "#/components/schemas/PageInfo"
          },
          "touchpoints" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Touchpoint"
            }
          }
        }
      },
...
    },
    "securitySchemes" : {
      "JWT" : {
        "bearerFormat" : "JWT",
        "description" : "JWT token get after Login",
        "scheme" : "bearer",
        "type" : "http"
      },
      "SubKey" : {
        "description" : "The Azure Subscription Key to access this API.",
        "in" : "header",
        "name" : "Ocp-Apim-Subscription-Key",
        "type" : "apiKey"
      }
    }
  }
}